fix query underlining for course descriptions
